extract sortHand function form bidding phase to cards entity
