+++ src/game/entities/cards.ts
@@ -1,4 +1,5 @@
 import flatten from "lodash/flatten";
+import groupBy from "lodash/groupBy";
 import { PlayerID } from "./players";
 
 export enum Suit {
@@ -189,4 +190,29 @@ export function generateCardDeck(ranks: Rank[] = allRanks): Card[] {
   );
 
   return cards;
+}
+
+export function sortHand(hand: Card[], trumpSuit?: Suit | null): Card[] {
+  // group cards by suit / N / Z
+  const groupedHand = groupBy(hand, (card) => {
+    if (card.rank === Rank.Z || card.rank === Rank.N) return card.rank;
+    return card.suit;
+  });
+
+  const sortedGroups = Object.entries(groupedHand)
+    // sort groups in this order: [Ns, ... non-trump suits sorted by num card, trump suit, Zs]
+    .sort(([keyA, cardsA], [keyB, cardsB]) => {
+      // always sort N left and Z right
+      if (keyA === Rank.N.toString() || keyB === Rank.Z.toString()) return -1;
+      if (keyA === Rank.Z.toString() || keyB === Rank.N.toString()) return 1;
+      // always sort trump beside Z
+      if (keyA === trumpSuit) return 1;
+      if (keyB === trumpSuit) return -1;
+      // sort by number of cards of color
+      return cardsA.length - cardsB.length;
+    })
+    // sort cards in each group by rank
+    .map(([, cardGroup]) => cardGroup.sort((a, b) => a.rank - b.rank));
+  // set sorted hand to state
+  return flatten(sortedGroups);
 }
